requirements.txt
# ============================================
# PRODUCTION-GRADE ML STACK FOR CHURN PREDICTION
# ============================================

# Core ML Libraries (Gradient Boosting - NOT Deep Learning!)
xgboost>=2.0.0
catboost>=1.2.0
lightgbm>=4.0.0
scikit-learn>=1.3.0
imbalanced-learn>=0.11.0

# Data Manipulation
pandas>=2.0.0
numpy>=1.24.0
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# MLOps Stack
mlflow>=2.9.0
dvc>=3.30.0
dvc-s3>=3.0.0
evidently>=0.4.0

# Model Interpretability
shap>=0.44.0

# Visualization & EDA
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
# sweetviz>=2.3.0  # Not compatible with Python 3.13
# ydata-profiling>=4.0.0  # Renamed from pandas-profiling, optional for EDA


# API & Deployment
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
gunicorn>=21.2.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Frontend
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0.0
tqdm>=4.66.0
kaggle>=1.5.16
openpyxl>=3.1.0
python-box>=7.0.0
ensure>=1.0.0

# Testing (Optional but Professional)
pytest>=7.4.0
pytest-cov>=4.1.0

# Hyperparameter Tuning
optuna>=3.4.0

# Logging
loguru>=0.7.0

-e .

